+++ src/templates/displays/rooms.py
@@ -8,9 +8,9 @@ from kivy.lang import Builder
 from kivy.uix.label import Label
 
 from ...helpers.serializers import signed_timestamp
-from ..templates.layouts import Layouts
+from ..behaviors.layouts import Layouts
 from ..displays.message import MessageUI
-from ..templates.textinputs import InputFieldInterface
+from ..behaviors.textinputs import InputFieldInterface
 
 Builder.load_string("""
 <Room>:
