--- src/views/app.py
@@ -1,7 +1,0 @@
-from kivy.app import App
-
-from .dashboard import Dashboard
-
-class BluechClientApp(App):
-    def build(self):
-        return Dashboard()
